@@ -1,51 +1,33 @@
 {
   "name": "krystalconsiders",
   "version": "1.0.0",
-  "description": "Krystal&#39;s blog",
-  "author": "Timothy Stapleton",
+  "description": "Krystal's Blog Frontend",
+  "author": "Timothy Stapleton <retrocede@outlook.com>",
   "private": true,
   "scripts": {
     "dev": "nuxt",
     "build": "nuxt build",
     "start": "nuxt start",
-    "generate": "nuxt generate",
-    "lint": "eslint --ext .js,.vue --ignore-path .gitignore .",
-    "precommit": "npm run lint",
-    "test": "jest"
+    "generate": "nuxt generate"
   },
   "dependencies": {
-    "@nuxtjs/axios": "^5.3.6",
-    "@nuxtjs/markdownit": "^1.2.3",
-    "@nuxtjs/pwa": "^2.6.0",
-    "@sanity/client": "^0.140.11",
-    "cross-env": "^5.2.0",
-    "nuxt": "^2.4.0",
-    "nuxt-buefy": "^0.3.2",
-    "nuxtent": "^1.4.1"
+    "@tryghost/content-api": "^1.2.0",
+    "nuxt": "^2.2.0",
+    "vuetify": "^1.3.6",
+    "vuetify-loader": "^1.0.5"
   },
   "devDependencies": {
-    "@nuxtjs/eslint-config": "^0.0.1",
-    "@vue/test-utils": "^1.0.0-beta.27",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "^8.2.1",
-    "babel-jest": "^24.1.0",
-    "eslint": "^5.0.1",
-    "eslint-config-prettier": "^3.1.0",
-    "eslint-config-standard": ">=12.0.0",
-    "eslint-loader": "^2.0.0",
-    "eslint-plugin-import": ">=2.14.0",
-    "eslint-plugin-jest": ">=21.24.1",
-    "eslint-plugin-node": ">=7.0.1",
-    "eslint-plugin-nuxt": ">=0.4.2",
-    "eslint-plugin-prettier": "2.6.2",
-    "eslint-plugin-promise": ">=4.0.1",
-    "eslint-plugin-standard": ">=4.0.0",
-    "eslint-plugin-vue": "^5.0.0",
-    "jest": "^24.1.0",
-    "netlify-cms": "^2.6.1",
-    "nodemon": "^1.18.9",
-    "nuxt-netlify-cms": "^4.0.0",
-    "prettier": "1.14.3",
-    "vue-jest": "^3.0.3"
+    "babel-eslint": "^10.0.1",
+    "babel-plugin-transform-imports": "^1.5.1",
+    "eslint": "^5.7.0",
+    "eslint-config-standard": "^12.0.0",
+    "eslint-loader": "^2.1.1",
+    "eslint-plugin-html": "^4.0.6",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-node": "^7.0.1",
+    "eslint-plugin-promise": "^4.0.1",
+    "eslint-plugin-standard": "^4.0.0",
+    "stylus": "^0.54.5",
+    "stylus-loader": "^3.0.2"
   }
 }
