@@ -1,71 +1,60 @@
-const pkg = require('./package')
-const glob = require('glob')
-const path = require('path')
+import GhostContentAPI from '@tryghost/content-api'
+const nodeExternals = require('webpack-node-externals')
+const VuetifyLoaderPlugin = require('vuetify-loader/lib/plugin')
+const api = new GhostContentAPI({
+  url: 'https://ghost.krystalconsiders.com',
+  key: '76a2b608a386617f265fec1a3b',
+  version: 'v2'
+})
+
+function generateRoutes () {
+  return api.posts
+    .browse()
+    .then(posts => {
+      return posts.map(post => {
+        return `/${post.slug}`;
+      })
+    })
+}
 
 module.exports = {
-  mode: 'universal',
-
   /*
   ** Headers of the page
   */
   head: {
-    title: pkg.name,
+    title: 'krystalconsiders',
     meta: [
       { charset: 'utf-8' },
       { name: 'viewport', content: 'width=device-width, initial-scale=1' },
-      { hid: 'description', name: 'description', content: pkg.description }
+      { hid: 'description', name: 'description', content: 'Krystals Blog Frontend' }
     ],
-    link: [{ rel: 'icon', type: 'image/x-icon', href: '/favicon.ico' }]
+    link: [
+      { rel: 'icon', type: 'image/x-icon', href: '/favicon.ico' },
+      { rel: 'stylesheet', href: 'https://fonts.googleapis.com/css?family=Roboto:300,400,500,700|Material+Icons' }
+    ]
   },
-
+  plugins: ['~/plugins/vuetify.js'],
+  css: ['~/assets/style/app.styl'],
   /*
-  ** Customize the progress-bar color
+  ** Customize the progress bar color
   */
-  loading: { color: '#fff' },
-
+  loading: { color: '#3B8070' },
   /*
-  ** Global CSS
+  ** Generate configuration
   */
-  css: [],
-
-  /*
-  ** Plugins to load before mounting the App
-  */
-  plugins: [],
-
-  /*
-  ** Nuxt.js modules
-  */
-  modules: [
-    // Doc: https://axios.nuxtjs.org/usage
-    '@nuxtjs/axios',
-    // Doc: https://buefy.github.io/#/documentation
-    'nuxt-buefy',
-    '@nuxtjs/pwa',
-    '@nuxtjs/markdownit'
-  ],
-  /*
-  ** Axios module configuration
-  */
-  axios: {
-    // See https://github.com/nuxt-community/axios-module#options
+  generate: {
+    routes: generateRoutes
   },
-
-  markdownit: {
-    injected: true,
-    preset: 'default',
-    breaks: true,
-    html: true
-  },
-
   /*
   ** Build configuration
   */
   build: {
-    /*
-    ** You can extend webpack config here
-    */
-    extend(config, ctx) {
+    transpile: [/^vuetify/],
+    plugins: [
+      new VuetifyLoaderPlugin()
+    ],
+    extractCSS: true,
+    extend (config, ctx) {
       // Run ESLint on save
       if (ctx.isDev && ctx.isClient) {
         config.module.rules.push({
@@ -74,6 +63,13 @@ module.exports = {
           loader: 'eslint-loader',
           exclude: /(node_modules)/
         })
+      }
+      if (process.server) {
+        config.externals = [
+          nodeExternals({
+            whitelist: [/^vuetify/]
+          })
+        ]
       }
     }
   }
