@@ -1,20 +1,16 @@
 module.exports = {
   root: true,
+  parser: 'babel-eslint',
   env: {
     browser: true,
     node: true
   },
-  parserOptions: {
-    parser: 'babel-eslint'
-  },
-  extends: [
-    '@nuxtjs',
-    'plugin:nuxt/recommended',
-    'plugin:prettier/recommended'
-  ],
+  extends: 'standard',
+  // required to lint *.vue files
   plugins: [
-    'prettier'
+    'html'
   ],
   // add your custom rules here
-  rules: {}
+  rules: {},
+  globals: {}
 }
