@@ -1,9 +1,9 @@
 <template>
   <div class="VueToNuxtLogo">
-    <div class="Triangle Triangle--two" />
-    <div class="Triangle Triangle--one" />
-    <div class="Triangle Triangle--three" />
-    <div class="Triangle Triangle--four" />
+    <div class="Triangle Triangle--two"></div>
+    <div class="Triangle Triangle--one"></div>
+    <div class="Triangle Triangle--three"></div>
+    <div class="Triangle Triangle--four"></div>
   </div>
 </template>
 
@@ -29,7 +29,7 @@
 .Triangle--one {
   border-left: 105px solid transparent;
   border-right: 105px solid transparent;
-  border-bottom: 180px solid #41b883;
+  border-bottom: 180px solid #41B883;
 }
 
 .Triangle--two {
@@ -38,7 +38,7 @@
   animation: goright 0.5s linear forwards 3.5s;
   border-left: 87.5px solid transparent;
   border-right: 87.5px solid transparent;
-  border-bottom: 150px solid #3b8070;
+  border-bottom: 150px solid #3B8070;
 }
 
 .Triangle--three {
@@ -47,7 +47,7 @@
   animation: goright 0.5s linear forwards 3.5s;
   border-left: 70px solid transparent;
   border-right: 70px solid transparent;
-  border-bottom: 120px solid #35495e;
+  border-bottom: 120px solid #35495E;
 }
 
 .Triangle--four {
