@@ -1,13 +1,53 @@
 <template>
-  <section class="section">
-    <h2 class="title is-3 has-text-grey">
-      "Just start  <b-icon
-        icon="rocket"
-        size="is-large"
-      />"
-    </h2>
-    <h3 class="subtitle is-6 has-text-grey">
-      Author: <a href="https://github.com/anteriovieira">Antério Vieira</a>
-    </h3>
-  </section>
+  <v-layout>
+    <v-flex text-xs-center>
+      <img src="/v.png" alt="Vuetify.js" class="mb-5">
+      <blockquote class="blockquote">
+        &#8220;First, solve the problem. Then, write the code.&#8221;
+        <footer>
+          <small>
+            <em>&mdash;John Johnson</em>
+          </small>
+        </footer>
+      </blockquote>
+      <div class="posts">
+        <ul>
+          <li v-for="post in posts" :key="post.id">
+            <router-link :to="`/${post.slug}`">{{ post.title }}</router-link>
+          </li>
+        </ul>
+      </div>
+    </v-flex>
+  </v-layout>
 </template>
+
+<script>
+import GhostContentAPI from '@tryghost/content-api'
+
+export default {
+  data () {
+    return {
+      posts: []
+    }
+  },
+  created () {
+    const api = new GhostContentAPI({
+      url: 'https://ghost.krystalconsiders.com',
+      key: '76a2b608a386617f265fec1a3b',
+      version: 'v2'
+    })
+
+    api.posts
+      .browse({ include: 'tags,authors' })
+      .then(posts => {
+        this.posts = posts
+        posts.forEach(post => {
+          console.log(post.title)
+        })
+      })
+      .catch(err => {
+        console.log('error: ', err)
+      })
+  }
+}
+</script>
